fix: relocate gr2 CLI tests to the gr2-cli package
r1 caught a real CI failure this PR's original commit missed: 23 cli_tests
panicked on CARGO_BIN_EXE_gr2 being unset, because Cargo only sets that env
var for [[bin]] targets the SAME package declares, and these tests lived in
root -- which, after this PR's own manifest fix, only owns gr/gitgrip.
Moved the 23 gr2-testing functions to gr2/tests/cli_tests.rs, the package
that actually owns the gr2 binary, and added its missing dev-dependencies
(assert_cmd, predicates, tempfile). Root keeps its 22 gr/gitgrip tests
unchanged; 22 + 23 = 45, same total as before.

Also fixes an unrelated clippy::question_mark lint in griptree.rs that was
blocking CI independent of this PR's diff -- unlike too_many_arguments,
that lint isn't in CI's allowlist.

Verified clean-room: cargo clean + cargo test --all (0 failures, both
cli_tests binaries at 22/22 and 23/23), cargo fmt --check, and cargo
clippy with CI's exact ci.yml invocation.
