+++ src/core/griptree.rs
@@ -214,10 +214,7 @@ impl GriptreePointer {
                 }
             }
 
-            match current.parent() {
-                Some(parent) => current = parent.to_path_buf(),
-                None => return None,
-            }
+            current = current.parent()?.to_path_buf();
         }
     }
 }
